Add erc20 indexing
